chore(solver): Remove previous no-overlap implementation
